Move test component into the spec file
